Methods moved from NamingAnalyzer into NamesFinder
- Refactor naming-related logic into `NamesFinder`

- Update all relevant analyzers to use `NamesFinder`

- Enhanced code clarity by renaming parameters
