+++ MiKo.Analyzer.Shared/Rules/Naming/MiKo_1509_ParametersWithStructuralDesignPatternSuffixAnalyzer.cs
@@ -3,6 +3,8 @@
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
 
+using MiKoSolutions.Analyzers.Linguistics;
+
 namespace MiKoSolutions.Analyzers.Rules.Naming
 {
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
@@ -22,12 +24,12 @@ namespace MiKoSolutions.Analyzers.Rules.Naming
                 return false;
             }
 
-            return IsNameForStructuralDesignPattern(symbol.Name);
+            return NamesFinder.IsNameForStructuralDesignPattern(symbol.Name);
         }
 
         protected override IEnumerable<Diagnostic> AnalyzeName(IParameterSymbol symbol, Compilation compilation)
         {
-            var betterName = FindBetterNameForStructuralDesignPattern(symbol.Name);
+            var betterName = NamesFinder.FindBetterNameForStructuralDesignPattern(symbol.Name);
 
             return new[] { Issue(symbol, betterName, CreateBetterNameProposal(betterName)) };
         }
